Enum mpl_token_vault::instruction::VaultInstruction

pub enum VaultInstruction {
    InitVault(InitVaultArgs),
    AddTokenToInactiveVault(AmountArgs),
    ActivateVault(NumberOfShareArgs),
    CombineVault,
    RedeemShares,
    WithdrawTokenFromSafetyDepositBox(AmountArgs),
    MintFractionalShares(NumberOfShareArgs),
    WithdrawSharesFromTreasury(NumberOfShareArgs),
    AddSharesToTreasury(NumberOfShareArgs),
    UpdateExternalPriceAccount(ExternalPriceAccount),
    SetAuthority,
}

Instructions supported by the Fraction program.

Variants

InitVault(InitVaultArgs)

Initialize a token vault, starts inactivate. Add tokens in subsequent instructions, then activate. 0. [writable] Initialized fractional share mint with 0 tokens in supply, authority on mint must be pda of program with seed [prefix, programid]

[writable] Initialized redeem treasury token account with 0 tokens in supply, owner of account must be pda of program like above

[writable] Initialized fraction treasury token account with 0 tokens in supply, owner of account must be pda of program like above

[writable] Uninitialized vault account

[] Authority on the vault

[] Pricing Lookup Address

[] Token program

[] Rent sysvar

Tuple Fields of InitVault

0: InitVaultArgs

AddTokenToInactiveVault(AmountArgs)

Add a token to a inactive token vault 0. [writable] Uninitialized safety deposit box account address (will be created and allocated by this endpoint) Address should be pda with seed of [PREFIX, vault_address, token_mint_address]

[writable] Initialized Token account

[writable] Initialized Token store account with authority of this program, this will get set on the safety deposit box

[writable] Initialized inactive fractionalized token vault

[signer] Authority on the vault

[signer] Payer

[signer] Transfer Authority to move desired token amount from token account to safety deposit

[] Token program

[] Rent sysvar

[] System account sysvar

Tuple Fields of AddTokenToInactiveVault

0: AmountArgs

ActivateVault(NumberOfShareArgs)

Activates the vault, distributing initial shares into the fraction treasury. Tokens can no longer be removed in this state until Combination. 0. [writable] Initialized inactivated fractionalized token vault

[writable] Fraction mint

[writable] Fraction treasury

[] Fraction mint authority for the program - seed of [PREFIX, program_id]

[signer] Authority on the vault

[] Token program

Tuple Fields of ActivateVault

0: NumberOfShareArgs

CombineVault

This act checks the external pricing oracle for permission to combine and the price of the circulating market cap to do so. If you can afford it, this amount is charged and placed into the redeem treasury for shareholders to redeem at a later time. The treasury then unlocks into Combine state and you can remove the tokens. 0. [writable] Initialized activated token vault

[writable] Token account containing your portion of the outstanding fraction shares

[writable] Token account of the redeem_treasury mint type that you will pay with

[writable] Fraction mint

[writable] Fraction treasury account

[writable] Redeem treasury account

[] New authority on the vault going forward - can be same authority if you want

[signer] Authority on the vault

[signer] Transfer authority for the token account and outstanding fractional shares account you’re transferring from

[] PDA-based Burn authority for the fraction treasury account containing the uncirculated shares seed [PREFIX, program_id]

[] External pricing lookup address

[] Token program

RedeemShares

If in the combine state, shareholders can hit this endpoint to burn shares in exchange for monies from the treasury. Once fractional supply is zero and all tokens have been removed this action will take vault to Deactivated 0. [writable] Initialized Token account containing your fractional shares

[writable] Initialized Destination token account where you wish your proceeds to arrive

[writable] Fraction mint

[writable] Redeem treasury account

[] PDA-based Transfer authority for the transfer of proceeds from redeem treasury to destination seed [PREFIX, program_id]

[signer] Burn authority for the burning of your shares

[] Combined token vault

[] Token program

[] Rent sysvar

WithdrawTokenFromSafetyDepositBox(AmountArgs)

If in combine state, authority on vault can hit this to withdrawal some of a token type from a safety deposit box. Once fractional supply is zero and all tokens have been removed this action will take vault to Deactivated 0. [writable] Initialized Destination account for the tokens being withdrawn

[writable] The safety deposit box account key for the tokens

[writable] The store key on the safety deposit box account

[writable] The initialized combined token vault

[] Fraction mint

[signer] Authority of vault

[] PDA-based Transfer authority to move the tokens from the store to the destination seed [PREFIX, program_id]

[] Token program

[] Rent sysvar

Tuple Fields of WithdrawTokenFromSafetyDepositBox

0: AmountArgs

MintFractionalShares(NumberOfShareArgs)

Self explanatory - mint more fractional shares if the vault is configured to allow such. 0. [writable] Fraction treasury

[writable] Fraction mint

[] The initialized active token vault

[] PDA-based Mint authority to mint tokens to treasury[PREFIX, program_id]

[signer] Authority of vault

[] Token program

Tuple Fields of MintFractionalShares

0: NumberOfShareArgs

WithdrawSharesFromTreasury(NumberOfShareArgs)

Withdraws shares from the treasury to a desired account. 0. [writable] Initialized Destination account for the shares being withdrawn

[writable] Fraction treasury

[] The initialized active token vault

[] PDA-based Transfer authority to move tokens from treasury to your destination[PREFIX, program_id]

[signer] Authority of vault

[] Token program

[] Rent sysvar

Tuple Fields of WithdrawSharesFromTreasury

0: NumberOfShareArgs

AddSharesToTreasury(NumberOfShareArgs)

Returns shares to the vault if you wish to remove them from circulation. 0. [writable] Initialized account from which shares will be withdrawn

[writable] Fraction treasury

[] The initialized active token vault

[signer] Transfer authority to move tokens from your account to treasury

[signer] Authority of vault

[] Token program

Tuple Fields of AddSharesToTreasury

0: NumberOfShareArgs

UpdateExternalPriceAccount(ExternalPriceAccount)

Helpful method that isn’t necessary to use for main users of the app, but allows one to create/update existing external price account fields if they are signers of this account. Useful for testing purposes, and the CLI makes use of it as well so that you can verify logic. 0. [writable] External price account

Tuple Fields of UpdateExternalPriceAccount

0: ExternalPriceAccount

SetAuthority

Sets the authority of the vault to a new authority.

[writable] Vault

[signer] Vault authority

[] New authority